requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic[email]==2.7.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
razorpay==1.4.2
motor==3.3.2
